external/CMSIS_5/CMSIS/CoreValidation/Project/requirements.txt
# -*- coding: utf-8 -*-
#
# Python requirements for build.py script
#
python-matrix-runner~=1.0
lxml~=4.8
